+++ packages/components/scripts/jobs/vue.js
@@ -1,0 +1,35 @@
+import fs from 'fs';
+import path from 'path';
+import * as jobs from './vue/index.js';
+import { createRunPrepare, runAdjustPackageVersion } from './shared.js';
+
+/**
+ * Run all steps to create new react components
+ * @param {String} settings.componentDistDir The absolute path to the component dist
+ * @param {String} settings.componentPackageDir The absolute path to the component root
+ * @param {String} settings.vuePackageDir The absolute path to the react package root
+ */
+export const runCreateVueWrappers = async ({
+  componentDistDir,
+  componentPackageDir,
+  vuePackageDir,
+}) => {
+  // Get the manifest information
+  const manifest = path.join(componentDistDir, '/custom-elements.json');
+  const metadata = JSON.parse(fs.readFileSync(manifest, 'utf8'));
+
+  // Internal vue package paths for usage in sub packages
+  const outDir = path.join(vuePackageDir, './src');
+  const distDir = path.join(vuePackageDir, './dist');
+  const componentDir = path.join(vuePackageDir, '/src/components');
+
+  await runAdjustPackageVersion('Vue: Adjusting vue package.json version field...')(componentPackageDir, vuePackageDir);
+  await createRunPrepare('Vue: Cleaning up artifacts...')(outDir, componentDir, distDir);
+  await jobs.runCreateWrappers(metadata, outDir);
+  await jobs.runFormat(outDir, vuePackageDir);
+
+  // @todo: This is currently disabled. We ship the vue files directly instead.
+  // Enabling this will be tackled in another ticket after v1.0.0
+  // await jobs.runEsBuild(distDir);
+  await jobs.runVueTypeScript(distDir, vuePackageDir);
+};