+++ packages/docs/.storybook/synergy-theme.ts
@@ -3,7 +3,7 @@ import { create } from '@storybook/theming/create';
 const baseConfig = {
   brandTitle: 'Synergy Design System',
   brandUrl: 'https://github.com/SickDesignSystem/synergy/',
-  brandImage: '/logo-claim.svg',
+  brandImage: '/synergy_logo_transparent.png',
   brandTarget: '_self',
 };