+++ packages/docs/stories/components/checkbox.stories.ts
@@ -1,81 +1,102 @@
-/* eslint-disable */
-
-/* eslint-disable import/no-relative-packages */
-
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
+import type { SynButton, SynCheckbox } from '@synergy-design-system/components';
 import '../../../components/src/components/checkbox/checkbox';
 import type { Meta, StoryObj } from '@storybook/web-components';
 import { html } from 'lit';
 import { userEvent } from '@storybook/testing-library';
-import { waitUntil } from '@open-wc/testing-helpers';
-import docsTokens from '../../../tokens/src/figma-tokens/_docs.json';
-import { storybookDefaults, storybookHelpers, storybookTemplate } from '../../src/helpers/component.js';
-const { args, argTypes } = storybookDefaults('syn-checkbox');
+import {
+  generateStoryDescription,
+  storybookDefaults,
+  storybookHelpers,
+  storybookTemplate,
+} from '../../src/helpers/component.js';
+
+const { argTypes } = storybookDefaults('syn-checkbox');
 const { overrideArgs } = storybookHelpers('syn-checkbox');
 const { generateTemplate } = storybookTemplate('syn-checkbox');
 
-const generateStoryDescription = (attributeName: string) => {
-  const story = (docsTokens?.components?.checkbox as Record<string, any>)?.[attributeName]?.description?.value ?? 'No Description';
-  return {
-    story,
-  }
-};
-
 const meta: Meta = {
-  component: 'checkbox',
   args: overrideArgs([
-    { name: 'default', type: 'slot', value: 'Checkbox' }
+    { name: 'default', type: 'slot', value: 'Checkbox' },
   ]),
   argTypes,
-  title: 'Components/syn-checkbox',
   parameters: {
     docs: {
-      description: generateStoryDescription('default'),
-    }
-  }
+      description: {
+        component: generateStoryDescription('checkbox', 'default'),
+      },
+    },
+  },
+  title: 'Components/syn-checkbox',
 };
 export default meta;
 
 type Story = StoryObj;
 
 export const Default = {
-  render: (args: any) => {
-    return generateTemplate({ args });
-  },
   parameters: {
     docs: {
-      description: generateStoryDescription('default'),
-    }
-  }
+      description: {
+        story: generateStoryDescription('checkbox', 'default'),
+      },
+    },
+  },
+  render: (args: unknown) => generateTemplate({ args }),
 } as Story;
 
 export const Checked: Story = {
-  render: () => html`<syn-checkbox checked>Checked</syn-checkbox>`,
   parameters: {
+    controls: {
+      disable: true,
+    },
     docs: {
-      description: generateStoryDescription('checked'),
-    }
-  }
+      description: {
+        story: generateStoryDescription('checkbox', 'checked'),
+      },
+    },
+  },
+  render: () => html`<syn-checkbox checked>Checked</syn-checkbox>`,
 };
 
 export const Indeterminate: Story = {
-  render: () => html`<syn-checkbox indeterminate>Indeterminate</syn-checkbox>`,
   parameters: {
+    controls: {
+      disable: true,
+    },
     docs: {
-      description: generateStoryDescription('indeterminate'),
-    }
-  }
+      description: {
+        story: generateStoryDescription('checkbox', 'indeterminate'),
+      },
+    },
+  },
+  render: () => html`<syn-checkbox indeterminate>Indeterminate</syn-checkbox>`,
 };
 
 export const Disabled: Story = {
-  render: () => html`<syn-checkbox disabled>Disabled</syn-checkbox>`,
   parameters: {
+    controls: {
+      disable: true,
+    },
     docs: {
-      description: generateStoryDescription('disabled'),
-    }
-  }
+      description: {
+        story: generateStoryDescription('checkbox', 'disabled'),
+      },
+    },
+  },
+  render: () => html`<syn-checkbox disabled>Disabled</syn-checkbox>`,
 };
 
 export const Sizes: Story = {
+  parameters: {
+    controls: {
+      disable: true,
+    },
+    docs: {
+      description: {
+        story: generateStoryDescription('checkbox', 'sizes'),
+      },
+    },
+  },
   render: () => html`
     <div style="display: flex; flex-direction: column; gap: 1rem;">
       <syn-checkbox size="small">Small</syn-checkbox>
@@ -83,22 +104,35 @@ export const Sizes: Story = {
       <syn-checkbox size="large">Large</syn-checkbox>
     </div>
   `,
-  parameters: {
-    docs: {
-      description: generateStoryDescription('sizes'),
-    }
-  }
 };
 
 export const CustomValidity: Story = {
+  parameters: {
+    controls: {
+      disable: true,
+    },
+    docs: {
+      description: {
+        story: generateStoryDescription('checkbox', 'validity'),
+      },
+    },
+  },
   play: async ({ canvasElement }) => {
     try {
-      const button = canvasElement.querySelector('button');
+      const form = canvasElement.querySelector('form');
+      const button = canvasElement.querySelector('syn-button') as SynButton;
+      const checkbox = canvasElement.querySelector('syn-checkbox') as SynCheckbox;
+      const errorMessage = "Don't forget to check me!";
 
-      await waitUntil(() => customElements.whenDefined('syn-checkbox'));
+      if (form && button && checkbox) {
+        checkbox.setCustomValidity(errorMessage);
 
-      if (button) {
+        // make sure to always fire both events:
+        // 1. userEvent.click is needed for storybooks play function to register
+        // 2. button.click is needed to really click the button
+        // userEvent.click works on native elements only
         await userEvent.click(button);
+        button.click();
       }
     } catch (error) {
       console.error('Error in play function:', error);
@@ -107,49 +141,31 @@ export const CustomValidity: Story = {
   render: () => html`
     <form class="custom-validity">
       <syn-checkbox name="checked" value="on">Check me</syn-checkbox>
-      <br />
-      <button type="submit">Submit</button>
+      <syn-button type="submit" variant="filled">Submit</syn-button>
     </form>
     <style>
-    form {
-      display: flex;
+    .custom-validity {
+      display: inline-flex;
       flex-direction: column;
-    }
-
-    button {
-      margin-top: 1rem;
-      align-self: flex-start;
-      padding: 0.5rem 1rem;
-      min-width: 5%;
+      gap: 1rem;
     }
     </style>
 
     <script type="module">
-      const form = document.querySelector('.custom-validity');
-      const checkbox = form.querySelector('syn-checkbox');
-      const errorMessage = \`Don't forget to check me!\`;
+    const form = document.querySelector('.custom-validity');
+    const checkbox = form.querySelector('syn-checkbox');
+    const errorMessage = "Don't forget to check me!";
 
-      // Set initial validity as soon as the element is defined
-      customElements.whenDefined('syn-checkbox').then(async () => {
-        await checkbox.updateComplete;
-        checkbox.setCustomValidity(errorMessage);
-      });
+    // Update validity on change
+    checkbox.addEventListener('syn-change', () => {
+      checkbox.setCustomValidity(checkbox.checked ? '' : errorMessage);
+    });
 
-      // Update validity on change
-      checkbox.addEventListener('syn-change', () => {
-        checkbox.setCustomValidity(checkbox.checked ? '' : errorMessage);
-      });
-
-      // Handle submit
-      form.addEventListener('submit', event => {
-        event.preventDefault();
-        alert('All fields are valid!');
-      });
+    // Handle submit
+    form.addEventListener('submit', event => {
+      event.preventDefault();
+      alert('All fields are valid!');
+    });
     </script>
   `,
-  parameters: {
-    docs: {
-      description: generateStoryDescription('validity'),
-    }
-  },
 };