+++ packages/docs/stories/components/radio.stories.ts
@@ -1,28 +1,26 @@
+/* eslint-disable @typescript-eslint/no-unsafe-argument */
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
 /* eslint-disable import/no-relative-packages */
 
 import '../../../components/src/components/radio/radio.js';
 import type { Meta, StoryObj } from '@storybook/web-components';
 import { html } from 'lit';
 import { userEvent } from '@storybook/testing-library';
-import { waitUntil } from '@open-wc/testing-helpers';
-import docsTokens from '../../../tokens/src/figma-tokens/_docs.json';
-import { storybookDefaults, storybookHelpers, storybookTemplate } from '../../src/helpers/component.js';
+import { generateStoryDescription, storybookDefaults, storybookHelpers, storybookTemplate } from '../../src/helpers/component.js';
 
 const { args, argTypes } = storybookDefaults('syn-radio');
 const { overrideArgs } = storybookHelpers('syn-radio');
 const { generateTemplate } = storybookTemplate('syn-radio');
 
-const generateStoryDescription = (attributeName: string) => ({
-  story: (docsTokens?.components?.radio as Record<string, unknown>)?.[attributeName]?.description?.value ?? 'No Description',
-});
-
 const meta: Meta = {
-  component: 'radio',
-  args,
+  args: overrideArgs({ name: 'default', type: 'slot', value: 'Option' }, args),
   argTypes,
+  component: 'radio',
   parameters: {
     docs: {
-      description: generateStoryDescription('default'),
+      description: {
+        story: generateStoryDescription('radio', 'default'),
+      },
     },
   },
   title: 'Components/syn-radio',
@@ -34,17 +32,25 @@ type Story = StoryObj;
 export const Default = {
   parameters: {
     docs: {
-      description: generateStoryDescription('default'),
+      description: {
+        story: generateStoryDescription('radio', 'default'),
+      },
     },
   },
-  render: () => html`
-  <syn-radio value="1" size="medium">Option</syn-radio>`,
+  render: (storyArgs: unknown) => generateTemplate({ args: storyArgs }),
+  // render: () => html`
+  // <syn-radio value="1" size="medium">Option</syn-radio>`,
 } as Story;
 
 export const Disabled: Story = {
   parameters: {
+    controls: {
+      disable: true,
+    },
     docs: {
-      description: generateStoryDescription('disabled'),
+      description: {
+        story: generateStoryDescription('radio', 'disabled'),
+      },
     },
   },
   render: () => html`
@@ -53,8 +59,13 @@ export const Disabled: Story = {
 
 export const Focus: Story = {
   parameters: {
+    controls: {
+      disable: true,
+    },
     docs: {
-      description: generateStoryDescription('focus'),
+      description: {
+        story: generateStoryDescription('radio', 'focus'),
+      },
     },
   },
   play: ({ canvasElement }: { canvasElement: HTMLElement }) => {
@@ -69,51 +80,60 @@ export const Focus: Story = {
 
 export const Invalid: Story = {
   parameters: {
+    controls: {
+      disable: true,
+    },
     docs: {
-      description: generateStoryDescription('invalid'),
+      description: {
+        story: generateStoryDescription('radio', 'invalid'),
+      },
     },
   },
   play: async ({ canvasElement }: { canvasElement: HTMLElement }) => {
     try {
-      const radioGroup = canvasElement.querySelector('syn-radio-group');
       const radio = canvasElement.querySelector('syn-radio');
-      const button = canvasElement.querySelector('button');
+      const button = canvasElement.querySelector('syn-button');
 
-      await waitUntil(() => radioGroup?.shadowRoot?.querySelector('fieldset'));
-
-      if (button instanceof HTMLButtonElement && radio) {
+      if (button && radio) {
+        // make sure to always fire both events:
+        // 1. userEvent.click is needed for storybooks play function to register
+        // 2. button.click is needed to really click the button
+        // userEvent.click works on native elements only
         await userEvent.click(button);
+        button.click();
       }
     } catch (error) {
       console.error('Error in play function:', error);
     }
   },
   render: () => html`
-  <form>
+  <form class="custom-validity">
     <syn-radio-group required>
       <syn-radio value="1">Option</syn-radio>
     </syn-radio-group>
-    <button size="medium" type="submit">Submit</button>
+    <syn-button type="submit" variant="filled">Submit</syn-button>
   </form>
   <style>
-    form {
+  .custom-validity {
     display: flex;
     flex-direction: column;
-    }
-
-    button {
-      margin-top: 1rem;
-      align-self: flex-end;
-      padding: 0.5rem 1rem;
-      min-width: 5%;
-    }
+    gap: 1rem;
+  }
+  syn-button {
+    align-self: flex-start;
+  }
   </style>`,
 };
 
 export const Sizes: Story = {
   parameters: {
+    controls: {
+      disable: true,
+    },
     docs: {
-      description: generateStoryDescription('sizes'),
+      description: {
+        story: generateStoryDescription('radio', 'sizes'),
+      },
     },
   },
   render: () => html`